+++ experimental/cognee-adapter/src/stigmem_plugin_cognee/adapter.py
@@ -4,7 +4,7 @@ Bridge between stigmem's atomic fact store and Cognee's knowledge-graph layer.
 
 Two surfaces::
 
-    from adapter import StigmemCogneeAdapter
+    from stigmem_plugin_cognee.adapter import StigmemCogneeAdapter
 
     bridge = StigmemCogneeAdapter.from_env()
 
@@ -177,7 +177,7 @@ class StigmemCogneeAdapter:
         self._default_dataset = default_dataset
 
     @classmethod
-    def from_env(cls) -> "StigmemCogneeAdapter":
+    def from_env(cls) -> StigmemCogneeAdapter:
         """Build adapter from environment variables.
 
         Cognee configuration variables read here::